// src/models/powers/power-list.enum.ts
export enum COMBAT_POWERS {
  ACUIDADE_COM_ARMA = 'Acuidade com Arma',
  ARMA_SECUNDARIA_GRANE = 'Arma Secundária Grande',
  ARREMESSO_POTENTE = 'Arremesso Potente',
  ATAQUE_PESADO = 'Ataque Pesado',
  ATAQUE_POEROSO = 'Ataque Poderoso',
  ATAQUE_PRECISO = 'Ataque Preciso',
  BLOQUEIO_COM_ESCUDO = 'Bloqueio com Escudo',
  CARA_DE_CAVALARIA = 'Carga de Cavalaria',
  COMBATE_DEFENSIVO = 'Combate Defensivo',
  DERRUBAR_APRIMORADO = 'Derrubar Aprimorado',
  DESARMAR_APRIMORADO = 'Desarmar Aprimorado',
  DISPARO_RAPIDO = 'Disparo Rápido',
  DISPARO_PRECISO = 'Disparo Preciso',
  EMPUNHADURA_PODEROSA = 'Empunhadura Poderosa',
  ENCOURACADO = 'Encouraçado',
  ESQUIVA = 'Esquiva',
  ESTILO_DE_ARMA_E_ESCUDO = 'Estilo de Arma e Escudo',
  ESTILO_DE_ARREMESSO = 'Estilo de Arremesso',
  ESTILO_DE_DISPARO = 'Estilo de Disparo',
  ESTILO_DE_DUAS_ARMAS = 'Estilo de Duas Armas',
  ESTILO_DE_DUAS_MAOS = 'Estilo de Duas Mãos',
  ESTILO_DE_UMA_ARMA = 'Estilo de Uma Arma',
  ESTILO_DESARMADO = 'Estilo Desarmado',
  FANATICO = 'Fanático',
  FINTA_APRIMORADA = 'Finta Aprimorada',
  FOCO_EM_ARMA = 'Foco em Arma',
  GINETE = 'Ginete',
  INEXPUGNAVEL = 'Inexpugnável',
  MIRA_APURADA = 'Mira Apurada',
  PRESENCA_ATERRADORA = 'Presença Aterradora',
  PROFICIENCIA_ARMAS_MARCIAIS = 'Proficiência em Armas Marciais',
  PROFICIENCIA_ARMAS_DE_FOGO = 'Proficiência em Armas de Fogo',
  PROFICIENCIA_ARMADURAS_PEDADAS = 'Proficiência em Armaduras Pesadas',
  PROFICIENCIA_ESCUDO = 'Proficiência em Escudos',
  PROFICIENCIA_ARMAS_EXOTICAS = 'Proficiência em Armas Exóticas',
  QUEBRAR_APRIMORADO = 'Quebrar Aprimorado',
  REFLEXOS_DE_COMBATE = 'Reflexos de Combate',
  SAQUE_RAPIDO = 'Saque Rápido',
  TRESPASSAR = 'Trespassar',
  VITALIDADE = 'Vitalidade'
}

export enum DESTINY_POWERS {
  ACROBATICO = 'Acrobático',
  AO_SABOR_DO_DESTINO = 'Ao Sabor do Destino',
  APARENCIA_INOFENCISA = 'Aparência Inofensiva',
  ATLETICO = 'Atlético',
  ATRAENTE = 'Atraente',
  COMANDAR = 'Comandar',
  FOCO_EM_PERICIA = 'Foco em Perícia',
  INVESTIGADOR = 'Investigador',
  LOBO_SOLITARIO = 'Lobo Solitário',
  MEDICINA = 'Medicina',
  PARCEIRO = 'Parceiro',
  SENTIDOS_AGUCADOS = 'Sentidos Aguçados',
  SORTUDO = 'Sortudo',
  SURTO_HEROICO = 'Surto Heróico',
  TORCIDA = 'Torcida',
  TREINAMENTO_EM_PERICIA = 'Treinamento em Perícia',
  VENEFICIO = 'Venefício',
  VONTADE_DE_FERRO = 'Vontade de Ferro'
}

export enum MAGIC_POWERS {
  CELEBRAR_RITUAL = 'Celebrar Ritual',
  ESCREVER_PERGAMINHO = 'Escrever Pergaminho',
  FOCO_EM_MAGIA = 'Foco em Magia',
  MAGIA_ACELERADA = 'Magia Acelerada',
  MAGIA_AMPLIADA = 'Magia Ampliada',
  MAGIA_DISCRETA = 'Magia Discreta',
  MAGIA_LIMITADA = 'Magia Limitada',
  PREPARAR_POCAO = 'Preparar Poção'
}

export enum GRANTED_POWERS {
  AFINIDADE_COM_A_TORMENTA = 'Afinidade com a Tormenta',
  ANFIBIO = 'Anfíbio',
  ARMAS_DA_AMBICAO = 'Armas da Ambição',
  ARSENAL_DAS_PROFUNDEZAS = 'Arsenal das Profundezas',
  ASTUCIA_DA_SERPENTE = 'Astúcia da Serpente',
  ATAQUE_PIEDOSO = 'Ataque Piedoso',
  AURA_DO_MEDO = 'Aura do Medo',
  AURA_DE_PAZ = 'Aura de Paz',
  AURA_RESTAURADORA = 'Aura Restauradora',
  BENCAO_DE_MANA = 'Benção de Mana',
  CARICIA_SOMBRIA = 'Carícia Sombria',
  CENTELHA_MAGICA = 'Centalha Mágica',
  CONHECIMENTO_ENCICLOPEDICO = 'Conhecimento Enciclopédico',
  CONJURAR_ARMA = 'Conjurar Arma',
  CORAGEM_TOTAL = 'Coragem Total',
  CURA_GENTIL = 'Cura Gentil',
  CURANDEIRA_PERFEITA = 'Curandeira Perfeita',
  DEDO_VERDE = 'Dedo Verde',
  DESCANSO_NATURAL = 'Descanso Natural',
  DOM_DA_IMORTALIDADE = 'Dom da Imortalidade',
  DOM_DA_PROFECIA = 'Dom da Profecia',
  DOM_DA_RESSUREICAO = 'Dom da Ressureição',
  DOM_DA_VERDADE = 'Dom da Verdade',
  ESCAMAS_DRACONICAS = 'Escamas Dracônicas',
  ESCUDO_MAGICO = 'Escudo Mágico',
  ESPADA_JUSTICEIRA = 'Espada Justiceira',
  ESPADA_SOLAR = 'Espada Solar',
  FARSA_DO_FINGIDOR = 'Farsa do Fingidor',
  FORMA_DE_MACACO = 'Forma de Macaco',
  FURIA_DIVINA = 'Furia Divina',
  GOLPISTA_DIVINO = 'Golpista Divino',
  HABITANTE_DO_DESERTO = 'Habitante do Deserto',
  INIMIGO_DE_TENEBRA = 'Inimigo de Tenebra',
  KIAI_DIVINO = 'Kiai Divino',
  LIBERDADE_DIVINA = 'Liberdade Divina',
  MANTO_DA_PENUMBRA = 'Manto da Penumbra',
  MENTE_ANALITICA = 'Mente Analítica',
  MENTE_VAZIA = 'Mente Vazia',
  MESTRE_DOS_MARES = 'Mestre dos Mares',
  OLHAR_AMEDRONTADOR = 'Olhar Amendrontador',
  PALAVRAS_DE_BONDADE = 'Palavras de Bondade',
  PERCEPCAO_TEMPORAL = 'Percepção Temporal',
  PODER_OCULTO = 'Poder Oculto',
  PRESAS_VENENOSAS = 'Presas Venenosas',
  REJEICAO_DIVINA = 'Rejeição Divina',
  SANGUE_DE_FERRRO = 'Sangue de Ferro',
  SANGUE_OFIDICO = 'Sangue Ofídico',
  SERVOS_DO_DRAGAO = 'Servos do Dragão',
  SORTE_DOS_LOUCOS = 'Sorte dos Loucos',
  TALENTO_ARTISTICO = 'Talento Artístico',
  TEURGISTA_MISTICO = 'Teurgista Místico',
  TRANSMISSAO_DA_LOUCURA = 'Transmissão da Loucura',
  TROPAS_DUYSHIDAKK = 'Tropas Duyshidakk',
  URRO_DIVINO = 'Urro Divino',
  VISAO_NAS_TREVAS = 'Visão nas Trevas',
  VOZ_DA_CIVILIZACAO = 'Voz da Civilização',
  VOZ_DA_NATUREZA = 'Voz da Natureza',
  VOZ_DOS_MONSTROS = 'Voz dos Monstros'
}

export enum TORMENT_POWERS {
  ANATOMIA_INSANA = 'Anatomia Insana',
  ANTENAS = 'Antenas',
  ARMAMENTO_ABERRANTE = 'Armamento Aberrante',
  ARTICULACOES_FLEXIVEIS = 'Articulações Flexíveis',
  ASAS_INSETOIDES = 'Asas Insetóides',
  CARAPACA = 'Carapaça',
  CORPO_ABERRANTE = 'Corpo Aberrante',
  DENTES_AFIADOS = 'Dentes Afiados',
  EMPUNHADURA_RUBRA = 'Empunhadura Rubra',
  MAOS_MEMBRANOSAS = 'Mãos Membranosas',
  MEMBROS_EXTRAS = 'Membros Extras',
  OLHOS_VERMELHOS = 'Olhos Vermelhos',
  PELE_CORROMPIDA = 'Pele Corrompida',
  SANGUE_ACIDO = 'Sangue Ácido',
  VISCO_RUBRO = 'Visco Rubro'
}